news/requirements.txt
flask
newsapi-python
webargs
Flask-JSON
